chore(kics): code review
